Add firebase hosting GitHub actions and remove trial level query results
